+++ cli/modify.c
@@ -16,12 +16,18 @@
 
 static void modify_usage(void)
 {
-	uprint("cobalt modify <id> @<board>\n");
+	uprint("cobalt modify <id> [--delete|-d] [--no-gc|-n]  [@<board>]\n");
 }
 
 static void modify_usage_long(void)
 {
 	modify_usage();
+	fprintf(stderr, "\n");
+	fprintf(stderr, "options:\n");
+	fprintf(stderr, "\t--delete|-d\n");
+	fprintf(stderr, "\t\tdelete the specified task\n");
+	fprintf(stderr, "\t--no-gc|-n\n");
+	fprintf(stderr, "\t\tdo not run garbage collection after deletion\n");
 	fprintf(stderr, "\n");
 	fprintf(stderr, "arguments:\n");
 	fprintf(stderr, "\t<id>\tnumeric id of the task to modify\n");
@@ -34,6 +40,7 @@ static int modify_main(int argc, const char *argv[])
 	struct cobalt *co;
 	int rc;
 	int err;
+	uint32_t flags;
 
 	rc = modify_parse(argc, argv, &cmd);
 	if (rc != 0) {
@@ -41,13 +48,13 @@ static int modify_main(int argc, const char *argv[])
 		return EXIT_FAILURE;
 	}
 
-	if (cmd.board == NULL) {
+	if (!cmd.del && cmd.board == NULL) {
 		eprint("board required when moving task\n");
 		return EXIT_FAILURE;
 	}
 
 	if (cmd.id == 0) {
-		eprint("id required when moving task\n");
+		eprint("id required when modifying task\n");
 		return EXIT_FAILURE;
 	}
 
@@ -57,19 +64,22 @@ static int modify_main(int argc, const char *argv[])
 		return EXIT_FAILURE;
 	}
 
-	rc = co_mod_attr(co, cmd.id, "board", cmd.board);
-
-	if (rc != CO_ENOERR) {
-		eprint("cannot move to board '%s': %s\n", cmd.board,
-				co_strerror(co));
-		rc = EXIT_FAILURE;
+	if (cmd.del) {
+		flags = cmd.nogc ? CO_DEL_NGC : 0;
+		rc = co_delete(co, cmd.id, flags);
+		if (rc != CO_ENOERR)
+			eprint("cannot delete task %08x: %s\n", cmd.id,
+					co_strerror(co));
 	} else {
-		rc = EXIT_SUCCESS;
+		rc = co_mod_attr(co, cmd.id, "board", cmd.board);
+		if (rc != CO_ENOERR)
+			eprint("cannot move to board '%s': %s\n", cmd.board,
+					co_strerror(co));
 	}
 
 	co_free(co);
 
-	return rc;
+	return rc == CO_ENOERR ? EXIT_SUCCESS : EXIT_FAILURE;
 }
 
 static struct module modify_module = {