+++ cli/show.h
@@ -6,10 +6,13 @@
 #define SHOW_H_
 
 
+#include <stdint.h>
+
 struct show_cmd {
 	const char *board;
 	const char *format;
 	int noboard;
+	uint32_t id;
 };
 
 int show_parse(int argc, const char *argv[], struct show_cmd *cmd);