+++ include/cobalt/cobalt.h
@@ -13,8 +13,8 @@
 
 /* major, minor, and patch version numbers */
 #define CO_MAJOR	0
-#define CO_MINOR	0
-#define CO_PATCH	1
+#define CO_MINOR	1
+#define CO_PATCH	0
 
 #define CO_API		__attribute__ ((visibility ("default")))