+++ CHANGES.md
@@ -6,6 +6,8 @@ All changes between releases will be documented in this file.
  - Man page for `cobalt` command using `asciidoc(1)`, along with `Makefile`
    changes to build documentation.
  - Added `printf(3)`-style format string argument to `show` command.
+ - Argument to the `add` command (`--data|-d`) may be used to specify data on
+   the command line, instead of opening `$EDITOR`.
 
 ### Fixed
  - Added newline to the end of `show` error message.